Began the tidyverse migration. This is by no means a working branch right now.

+++ phyloscannerR/R/tree_utility_functions.R
@@ -320,7 +320,7 @@ host.from.label <- function(label, regexp){
 
 read.count.from.label <- function(label, regexp){
   if(length(grep(regexp, label)>0)) {
-    return(as.numeric(sub(regexp, "\\3", label)))
+    return(as.integer(sub(regexp, "\\3", label)))
   } else {
     return(NA)
   }
